refactor ImdsCredential to use ManagedIdentityClient
